% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_stan_data.R
\name{assert_stan_data}
\alias{assert_stan_data}
\alias{assert_daily_data}
\alias{assert_country_data}
\title{Assert the data formats are correct}
\usage{
assert_stan_data(x)

assert_daily_data(x)

assert_country_data(x)
}
\arguments{
\item{x}{an object to check}
}
\description{
Assert the data formats are correct
}
