% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_daily_data.R
\name{check_daily_data}
\alias{check_daily_data}
\title{Check Daily Data set}
\usage{
check_daily_data(x)
}
\arguments{
\item{x}{a daily_data data.frame to check for inconsistencies}
}
\description{
Check Daily Data set
}
