% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_to_dataframe.R
\name{covariates_to_dataframe}
\alias{covariates_to_dataframe}
\title{Convert the covariates data to a data.frame}
\usage{
covariates_to_dataframe(
  x,
  start_date = "2020-01-01",
  end_date = Sys.Date(),
  nrows = NULL
)
}
\arguments{
\item{x}{a data.frame with variables Country, Type and Date}

\item{start_date}{The starting date to use}

\item{end_date}{The end date to use}

\item{nrows}{The total number of days}
}
\description{
The function computes a data.frame from the [interventions.csv] file.
}
