% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_stan_data.R
\name{covid19_stan_data}
\alias{covid19_stan_data}
\title{Create input data to the COVID19 Stan model}
\usage{
covid19_stan_data(
  formula,
  daily_data,
  country_data,
  serial_interval,
  ecdf_time,
  N0 = 6,
  N2 = 90,
  verbose = TRUE
)
}
\arguments{
\item{formula}{a R formula for the effect of covariates in [daily_data] on Rt}

\item{daily_data}{a data.frame that contains variables 
by date, such as cases, deaths and date-specific covariates.}

\item{country_data}{a data.frame that contains variables 
by by country, such as [ifr] and [total_population]}

\item{serial_interval}{Pre-calculated SI using emprical data from Neil}

\item{ecdf_time}{ECDF for infection-to-death}

\item{N0}{days of observed data + # of days to forecast}

\item{N2}{number of days for which to impute infections}
}
\description{
Create input data to the COVID19 Stan model
}
\details{
[daily_data] must include the following variables: 
* \code{date}
* \code{country}
* \code{cases} (the number of cases for that country and date)
* \code{deaths} (the number of deaths for that country and date)

[country_data] must include the following variables: 
* \code{country} (the same countries as in [daily_data])
* \code{ifr} (ifr per country)
* \code{total_population}
}
