% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_covidtracker_csv.R
\name{covidtracker_csv_v4}
\alias{covidtracker_csv_v4}
\title{Create and store API data as CSV files}
\usage{
covidtracker_csv_v4(countries = NULL, parse_factors = FALSE)
}
\arguments{
\item{countries}{Countrise to extract, \code{NULL} means all countries.}

\item{parse_factors}{pars ordinal variables to factors}
}
\description{
Create and store API data as CSV files
}
\details{
[covidtracker_csv_v4()] access and the published CSV-file directly as version 4
[covidtracker_csv_v5()] access and the published CSV-file directly as version 5
}
