% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_stan_data.R
\name{get_dates}
\alias{get_dates}
\title{Access dates for the epidemic period by country}
\usage{
get_dates(daily_data)
}
\arguments{
\item{daily_data}{a data.frame that contains variables
by date, such as cases, deaths and date-specific covariates.}
}
\description{
Access dates for the epidemic period by country
}
