% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecdc_data.R
\name{get_ecdc_data}
\alias{get_ecdc_data}
\title{Access and parse ECDC data}
\usage{
get_ecdc_data()
}
\description{
Access and parse ECDC data
}
\references{
\url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}
}
