% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_three_panel.R
\name{plot_deaths}
\alias{plot_deaths}
\title{Plot Daily Number of Deaths}
\usage{
plot_deaths(country, stan_fit, stan_data, daily_data)

plot_deaths(country, stan_fit, stan_data, daily_data)

plot_deaths(country, stan_fit, stan_data, daily_data)
}
\arguments{
\item{stan_fit}{a stan_fit object with [alpha] parameter draws}

\item{stan_data}{a stan_data object with named design matrix array}
}
\description{
Plot Daily Number of Deaths

Plot Daily Number of Deaths

Plot Daily Number of Deaths
}
