% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_three_panel.R
\name{plot_three_panel}
\alias{plot_three_panel}
\title{Plot threeplot}
\usage{
plot_three_panel(
  country,
  stan_fit,
  stan_data,
  daily_data,
  plot_dates = NULL,
  title = NULL
)
}
\arguments{
\item{stan_fit}{a stan_fit object with [alpha] parameter draws}

\item{stan_data}{a stan_data object with named design matrix array}
}
\description{
Plot threeplot
}
