% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_stan_covariate_data.R
\name{probability_of_death_given_infection}
\alias{probability_of_death_given_infection}
\title{Compute probability of death at day i given infection}
\usage{
probability_of_death_given_infection(ifr, ecdf_time, N2)
}
\arguments{
\item{ifr}{The overall probability of dying given infection}

\item{ecdf_time}{Cumulative distribution over time}

\item{N2}{total length of f}
}
\description{
Compute probability of death at day i given infection
}
