% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-covariates.r
\name{process_covariates}
\alias{process_covariates}
\title{Process Covariates}
\usage{
process_covariates(countries, interventions, d, ifr.by.country, N2)
}
\arguments{
\item{countries}{Countries to use}

\item{interventions}{Interventions object returned by [read_interventions()].}

\item{d}{Observation data (see [read_obs_data()])}

\item{ifr.by.country}{IFR by country (see [read_ifr_data()])}

\item{N2}{Increase if you need more forecast}
}
\description{
Process Covariates
}
