% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.r
\name{read_obs_data}
\alias{read_obs_data}
\alias{read_ifr_data}
\alias{read_interventions}
\title{Read Data}
\usage{
read_obs_data(countries)

read_ifr_data()

read_interventions(countries)
}
\arguments{
\item{countries}{Countries to use include in data}
}
\description{
Read Data
}
\details{
The functions reads data from the data folder.
}
